% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{calc_direction}
\alias{calc_direction}
\title{Calculate Direction Feature}
\usage{
calc_direction(x, y, window_size = 16, sampling_rate = 1000)
}
\arguments{
\item{x}{Numeric vector of x coordinates for each gaze sample.}

\item{y}{Numeric vector of y coordinates for each gaze sample.}

\item{window_size}{Window size in millisecond to calculate the direction feature.}

\item{sampling_rate}{Sampling rate of the eye-tracker (in Hz).}
}
\value{
A vector containing direction feature for each sample.
}
\description{
Calculates the direction feature.
}
