% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{calc_direction_deviation}
\alias{calc_direction_deviation}
\title{Calculate Direction Deviation}
\usage{
calc_direction_deviation(
  x,
  y,
  t,
  window_size = 16,
  outer_window_size = 200,
  outer_step_size = 10,
  sampling_rate = 1000
)
}
\arguments{
\item{x}{Numeric vector of x coordinates for each gaze sample.}

\item{y}{Numeric vector of y coordinates for each gaze sample.}

\item{t}{Numeric vector of time for each gaze sample.}

\item{window_size}{Window size in millisecond to calculate the direction feature.}

\item{outer_window_size}{Window size in millisecond to calculate the outer direction.}

\item{outer_step_size}{Step size in millisecond to calculate the outer direction.}

\item{sampling_rate}{Sampling rate of the eye-tracker (in Hz).}
}
\value{
A vector containing the standard deviation of the direction deviation for each sample (NA for first t < outer_step_size) .
}
\description{
Calculates the direction deviation.
}
