% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{calc_theta}
\alias{calc_theta}
\title{Calculate Sample-to-sample Angle}
\usage{
calc_theta(x, y)
}
\arguments{
\item{x}{Numeric vector of x coordinates for each gaze sample.}

\item{y}{Numeric vector of y coordinates for each gaze sample.}
}
\value{
A vector containing the sample-to-sample angle (in radians) for each gaze sample.
}
\description{
Calculates the relative angle between subsequent gaze samples.
Uses the backward difference to compute the absolute angle between two samples and
the forward difference for the change in angle at each sample.
}
