% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{calc_windowed_diff}
\alias{calc_windowed_diff}
\title{Calculate Windowed Diff}
\usage{
calc_windowed_diff(x, y, window_size, sampling_rate = 1000)
}
\arguments{
\item{x}{Numeric vector of x coordinates for each gaze sample.}

\item{y}{Numeric vector of y coordinates for each gaze sample.}

\item{window_size}{Window size in millisecond to calculate the direction feature.}

\item{sampling_rate}{Sampling rate of the eye-tracker (in Hz).}
}
\value{
A list containing the x and y difference vectors.
}
\description{
Calculates the windowed difference for x and y.
}
