% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma2.R
\docType{class}
\name{gamma2}
\alias{gamma2}
\alias{gamma2-class}
\alias{gamma2,ANY-method}
\title{Gamma Response Model for depmixS4}
\usage{
gamma2(y, pstart = NULL, fixed = NULL, ...)

\S4method{gamma2}{ANY}(y, pstart = NULL, fixed = NULL, ...)
}
\arguments{
\item{y}{Numeric vector containing a response variable.}
}
\value{
An S4 object of class 'gamma2'.
}
\description{
Creates a gamma response model to be used in \code{\link[depmixS4]{makeDepmix}}
as part of \code{\link{HMM_classify}}.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of length two containing the shape and scale parameters
of the gamma distribution.}

\item{\code{fixed}}{Logical vector indicating which parameters are fixed.}

\item{\code{y}}{Response variable.}

\item{\code{x}}{Covariate variables (or just an intercept).}

\item{\code{npar}}{Number of parameters.}

\item{\code{constr}}{Vector defining the parameter constraints.}
}}

