% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_start.R
\name{gamma_start}
\alias{gamma_start}
\title{Generate Gamma Distribution Starting Values}
\usage{
gamma_start(x, alpha = 3, seed)
}
\arguments{
\item{x}{A single numeric parameter value.}

\item{alpha}{Shape parameter of the gamma distribution.}

\item{seed}{Seed of the random number generation (for reproducibility).}
}
\value{
A single numeric starting value for the parameter.
}
\description{
Generates starting values from a gamma distribution to estimate a parameter. The mode of the distribution
is on the parameter value.
}
