% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gazeHMM.R
\name{gazeHMM}
\alias{gazeHMM}
\title{Classify Gaze Data into Eye Movement Events}
\usage{
gazeHMM(
  x,
  y,
  t,
  unit = "px",
  res,
  dim,
  dist,
  fr,
  blink = NULL,
  b.win = 0.05,
  sg.order = 3,
  sg.length = 5,
  nstates,
  respstart,
  trstart,
  instart,
  sf = c(10, 10),
  random.respstart = T,
  start.seed = NULL,
  fit.control = em.control(maxit = 5000, random.start = F),
  min.sac = 0.01
)
}
\arguments{
\item{x}{Numeric vector of raw x coordinates for each gaze samples.}

\item{y}{Numeric vector of raw y coordinates for each gaze samples.}

\item{t}{Numeric vector of time stamps for each gaze sample.}

\item{unit}{Character string indicating the unit of coordinates (either 'px' or 'va').}

\item{res}{Screen resolution (in px).}

\item{dim}{Screen dimensions (in mm).}

\item{dist}{Distance between subject and screen (in mm).}

\item{fr}{Sampling rate of the eye-tracker (in Hz).}

\item{blink}{Either a numeric vector of length two indicating the x and y coordinates for blink samples
or a logical vector with the same length as \code{x}, \code{y}, and \code{t} indicating blink samples.}

\item{b.win}{Time window around blink samples that are to be ignored (i.e., set to NA; in s).}

\item{sg.order}{Order of the Savitzky-Golay filter.}

\item{sg.length}{Length of the Savitzky-Golay filter (must be odd).}

\item{nstates}{Number of states in the hidden Markov model.}

\item{respstart}{Starting values for the response model in the hidden Markov model.}

\item{trstart}{Starting values for the transition model in the hidden Markov model.}

\item{instart}{Starting values for the initial state model in the hidden Markov model.}

\item{sf}{Vector of length two indicating by which factor velocity and acceleration data will be divided.}

\item{random.respstart}{Logical indicating whether random starting values will be generated for estimating
response parameters using \code{\link{gamma_start}}. Otherwise default or user-supplied starting values
will be used.}

\item{start.seed}{Seeds for generating random starting values for response parameters.}

\item{fit.control}{List of settings for the EM algorithm as returned by \code{\link[depmixS4]{em.control}}.}

\item{min.sac}{Minimum saccade duration (in s).}
}
\value{
A list object of class 'gazeHMM' containing four elements:
\describe{
  \item{samples}{Preprocessed gaze data with postprocessed sample labels.}
  \item{events}{A list with data frames for each classified event containing
  the event metrics (e.g., fixation duration).}
  \item{model}{The \code{depmix.fitted} model.}
  \item{settings}{The arguments of \code{gazeHMM} used to obtain results (for reproducibility).}
}
}
\description{
Classify Gaze Data into Eye Movement Events
}
