% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{HMM_classify}
\alias{HMM_classify}
\title{Classify Gaze Data}
\usage{
HMM_classify(
  data,
  nstates,
  respstart,
  trstart,
  instart,
  sf = c(10, 10),
  fit.control = em.control(maxit = 5000, random.start = F)
)
}
\arguments{
\item{data}{Data frame with preprocessed gaze samples as returned by \code{\link{preprocess}}.
The object must contain the variables \code{vel}, \code{acc}, and \code{angle}.}

\item{nstates}{Number of states in the hidden Markov model.}

\item{respstart}{Starting values for the response model in the hidden Markov model.}

\item{trstart}{Starting values for the transition model in the hidden Markov model.}

\item{instart}{Starting values for the initial state model in the hidden Markov model.}

\item{sf}{Vector of length two indicating by which factor velocity and acceleration data will be divided.}

\item{fit.control}{List of settings for the EM algorithm as returned by \code{\link[depmixS4]{em.control}}.}
}
\value{
An object of class \code{\link[depmixS4]{depmix.fitted}}.
}
\description{
Classifies gaze data into eye movement events by using a hidden Markov model from the package depmixS4.
}
