% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_helper_functions.R
\name{last_state}
\alias{last_state}
\title{Retrieve Last State}
\usage{
last_state(state, s)
}
\arguments{
\item{state}{Vector with estimated states for each gaze sample.}

\item{s}{Index of the current gaze sample in the state sequence.}
}
\value{
A single element indicating the label of the last different state.
}
\description{
Retrieves the last different state in a sequence of estimated states.
Used as an internal function in \code{\link{postprocess}}.
}
