% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{postprocess}
\alias{postprocess}
\title{Postprocess Gaze Data Classification}
\usage{
postprocess(data, state, fr, min.sac = 0.01)
}
\arguments{
\item{data}{Data frame with preprocessed gaze samples as returned by \code{\link{preprocess}}.
The object must contain the variables \code{t}, \code{vel}, \code{acc}, and \code{angle}.}

\item{state}{Vector of state labels for each sample as returned by \code{\link{HMM_classify}}.}

\item{fr}{Sampling rate of the eye-tracker (in Hz).}

\item{min.sac}{Minimum saccade duration (in s).}
}
\value{
A list with two objects: The preprocessed data with postprocessed sample labels and
a list with data frames for each classified event containing the event metrics (e.g., fixation duration).
}
\description{
Applies a postprocessing routine to classified gaze data as part of \code{\link{gazeHMM}}.
Relabels samples that violate constraints and computes eye movement event metrics.
}
