% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess Gaze Data}
\usage{
preprocess(
  x,
  y,
  t,
  unit = "px",
  res,
  dim,
  dist,
  fr,
  blink = NULL,
  b.win = 0.05,
  sg.order = 3,
  sg.length = sg.order + 3 - sg.order\%\%2
)
}
\arguments{
\item{x}{Numeric vector of raw x coordinates for each gaze samples.}

\item{y}{Numeric vector of raw y coordinates for each gaze samples.}

\item{t}{Numeric vector of time stamps for each gaze sample.}

\item{unit}{Character string indicating the unit of coordinates (either 'px' or 'va').}

\item{res}{Screen resolution (in px).}

\item{dim}{Screen dimensions (in mm).}

\item{dist}{Distance between subject and screen (in mm).}

\item{fr}{Sampling rate of the eye-tracker (in Hz).}

\item{blink}{Either a numeric vector of length two indicating the x and y coordinates for blink samples
or a logical vector with the same length as \code{x}, \code{y}, and \code{t} indicating blink samples.}

\item{b.win}{Time window around blink samples that are to be ignored (i.e., set to NA; in s).}

\item{sg.order}{Order of the Savitzky-Golay filter.}

\item{sg.length}{Length of the Savitzky-Golay filter (must be odd).}
}
\value{
A data frame with x and y gaze coordinates in degrees of visual angle, time stamps,
velocity, accelelration, sample-to-sample angle, and an initial label for each sample.
}
\description{
Prepares raw gaze data for classification by a hidden Markov model as part of \code{\link{gazeHMM}}.
}
