% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{px2va}
\alias{px2va}
\title{Convert Gaze Coordinates}
\usage{
px2va(x, res, dim, dist)
}
\arguments{
\item{x}{Numeric vector of raw coordinates for each gaze sample.}

\item{res}{Screen resolution (in px).}

\item{dim}{Screen dimensions (in mm).}

\item{dist}{Distance between subject and screen (in mm).}
}
\value{
A numeric vector with coordiantes in degrees of visual angle centered at (0,0) for each gaze sample.
}
\description{
Converts raw gaze coordinates from pixels into degrees of visual angle.
}
