% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_helper_functions.R
\name{reclassify}
\alias{reclassify}
\title{Relabel State Sequence}
\usage{
reclassify(state, t, min.sac)
}
\arguments{
\item{state}{Integer vector with estimated states for each gaze sample as
returned by \code{\link{HMM_classify}}.}

\item{t}{Numeric vector of time stamps for each gaze sample.}

\item{min.sac}{Minimum saccade duration (in s).}
}
\value{
A data frame with the relabeled states and event number for each gaze sample.
}
\description{
Relabels the state sequence estimated by a hidden Markov model with a recursive algorithm.
Used as an internal function in \code{\link{postprocess}}.
}
