% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-methods.R
\name{response-methods}
\alias{response-methods}
\alias{show,unif-method}
\alias{getpars,response-method}
\alias{setpars,unif-method}
\alias{predict,unif-method}
\alias{simulate,unif-method}
\alias{show,vMF-method}
\alias{setpars,vMF-method}
\alias{predict,vMF-method}
\alias{simulate,vMF-method}
\alias{show,gamma2-method}
\alias{setpars,gamma2-method}
\alias{predict,gamma2-method}
\alias{simulate,gamma2-method}
\title{Methods for Additional Response Models}
\usage{
\S4method{show}{unif}(object)

\S4method{getpars}{response}(object, which = "pars", ...)

\S4method{setpars}{unif}(object, values, which = "pars", ...)

\S4method{predict}{unif}(object)

\S4method{simulate}{unif}(object, nsim = 1, seed)

\S4method{show}{vMF}(object)

\S4method{getpars}{response}(object, which = "pars", ...)

\S4method{setpars}{vMF}(object, values, which = "pars", ...)

\S4method{predict}{vMF}(object)

\S4method{simulate}{vMF}(object, nsim = 1, seed)

\S4method{show}{gamma2}(object)

\S4method{getpars}{response}(object, which = "pars", ...)

\S4method{setpars}{gamma2}(object, values, which = "pars", ...)

\S4method{predict}{gamma2}(object)

\S4method{simulate}{gamma2}(object, nsim = 1, seed)
}
\arguments{
\item{object}{An object of class 'depmix', 'depmix.fitted', or 'response'.}

\item{which}{A single character element. Either 'pars', resulting in the
parameters of the object, or 'fixed', resulting in the logical vector of
which parameters are fixed.}

\item{...}{Further arguments to be passed to \code{setpars} or \code{getpars}}

\item{values}{A vector of values to be set as the parameters of the object.}

\item{nsim}{A single integer indicating how many times a data set should be
simulated from the object.}

\item{seed}{A single integer to determine the seed used for data simulation.}
}
\description{
Various methods for uniform, von Mises, and gamma response models to be used with \code{depmixS4}.
}
