% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif.R
\docType{class}
\name{unif}
\alias{unif}
\alias{unif-class}
\alias{unif,ANY-method}
\title{Uniform Response Model for depmixS4}
\usage{
unif(y, ...)

\S4method{unif}{ANY}(y, fixed = NULL, ...)
}
\arguments{
\item{y}{Numeric vector containing a response variable.}

\item{...}{Further arguments to be supplied by parent functions.}

\item{fixed}{Logical vector of length two indicating which parameters are
fixed.}
}
\value{
An S4 object of class 'unif'.
}
\description{
Creates a uniform response model to be used in \code{\link[depmixS4]{makeDepmix}}
as part of \code{\link{HMM_classify}}.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of length two containing the minimum and maximum parameters
of the uniform distribution.}

\item{\code{fixed}}{Logical vector indicating which parameters are fixed.}

\item{\code{y}}{Response variable.}

\item{\code{x}}{Covariate variables (or just an intercept).}

\item{\code{npar}}{Number of parameters.}

\item{\code{constr}}{Vector defining the parameter constraints.}
}}

