% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMF.R
\docType{class}
\name{vMF}
\alias{vMF}
\alias{vMF-class}
\alias{vMF,ANY-method}
\title{Von Mises Response Model for depmixS4}
\usage{
vMF(y, pstart = NULL, fixed = NULL, ...)

\S4method{vMF}{ANY}(y, pstart = NULL, fixed = NULL, ...)
}
\arguments{
\item{y}{Numeric vector containing a response variable.}

\item{pstart}{Numeric vector of length two containing the starting values
for the mean and concentration parameters.}

\item{fixed}{Logical vector of length two indicating which parameters are
fixed.}

\item{...}{Further arguments to be supplied by parent functions.}
}
\value{
An S4 object of class 'vMF'.
}
\description{
Creates a von Mises response model to be used in \code{\link[depmixS4]{makeDepmix}}
as part of \code{\link{HMM_classify}}.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of length two containing the mean and concentration parameters
of the von Mises distribution.}

\item{\code{fixed}}{Logical vector indicating which parameters are fixed.}

\item{\code{y}}{Response variable.}

\item{\code{x}}{Covariate variables (or just an intercept).}

\item{\code{npar}}{Number of parameters.}

\item{\code{constr}}{Vector defining the parameter constraints.}
}}

