% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-methods.R
\name{response-methods}
\alias{response-methods}
\alias{show,unif-method}
\alias{dens,unif-method}
\alias{getpars,response-method}
\alias{setpars,unif-method}
\alias{fit,unif-method}
\alias{predict,unif-method}
\alias{simulate,unif-method}
\alias{show,vMF-method}
\alias{dens,vMF-method}
\alias{setpars,vMF-method}
\alias{fit,vMF-method}
\alias{predict,vMF-method}
\alias{simulate,vMF-method}
\alias{show,gamma2-method}
\alias{dens,gamma2-method}
\alias{setpars,gamma2-method}
\alias{fit,gamma2-method}
\alias{predict,gamma2-method}
\alias{simulate,gamma2-method}
\title{Methods for Additional Response Models}
\usage{
\S4method{show}{unif}(object)

\S4method{dens}{unif}(object, log = FALSE)

\S4method{getpars}{response}(object, which = "pars", ...)

\S4method{setpars}{unif}(object, values, which = "pars", ...)

\S4method{fit}{unif}(object, w)

\S4method{predict}{unif}(object)

\S4method{simulate}{unif}(object, nsim = 1, seed)

\S4method{show}{vMF}(object)

\S4method{dens}{vMF}(object)

\S4method{getpars}{response}(object, which = "pars", ...)

\S4method{setpars}{vMF}(object, values, which = "pars", ...)

\S4method{fit}{vMF}(object, w)

\S4method{predict}{vMF}(object)

\S4method{simulate}{vMF}(object, nsim = 1, seed)

\S4method{show}{gamma2}(object)

\S4method{dens}{gamma2}(object, log = FALSE)

\S4method{getpars}{response}(object, which = "pars", ...)

\S4method{setpars}{gamma2}(object, values, which = "pars", ...)

\S4method{fit}{gamma2}(object, w)

\S4method{predict}{gamma2}(object)

\S4method{simulate}{gamma2}(object, nsim = 1, seed)
}
\description{
Various methods for uniform, von Mises, and gamma response models to be used with \code{depmixS4}.
}
\seealso{
\code{\link[depmixS4]{show}}.
}
